\name{rmf}
\alias{rmf}

\title{
  Remove Special Characters
}

\description{
  A simple helper function that removes special characters from a character string.
}

\usage{
rmf(x)
}

\arguments{
  \item{x}{A character string.}
}

\value{
  A character string with special characters removed.
}

\examples{
rmf("ba*&m^l$$:s.s")
}

\keyword{misc}

